Add upgrade/rollback test
If there is a newer image version to upgrade to, do that, ensure that we
upgraded to the expected latest version, roll back, and ensure that we are back
to where we started.

The test gets skipped if there is no newer version to upgrade to, or if it's
run without autopkgtest, or on an autopkgtest runner which doesn't offer
reboot.